updated eth-simple-keyring version
We need this feature for https://github.com/MetaMask/metamask-extension/issues/1190.
With this function, we can obtain the encryption public key for the user, that is generated via `nacl` library. Then we can store it and send encrypted messages to the user in a secure manner. The user will be able to decrypt them via "eth_decryptMessage"