Remove browser-related test dependencies
All test dependencies related to browser support have been removed.
This package works just fine in a Node.js environment - we don't need
JSDOM or anything like that. The only strict dependency we have upon
browser APIs is within the `browser-passworder` package, which uses
the `crypto` API (only available in Node.js >v15). This dependency is
used as the default encryptor, but we use a custom mock encryptor in
unit tests.