Add changelog
The package `@metamask/auto-changelog` has been added to help with
validating and updating the changelog. An initial empty changelog has
been added as well.

The `repository` property of `package.json` was updated to use the
HTTPS link to the git repo, because the `git+` link isn't compatible
with `@metamask/auto-changelog`.

Basic changelog validation has been added to CI.