Switch from mocha to Jest
We now use `jest` for testing, rather than `mocha`. The primary change
required to migrate the tests was to switch from `assert` to `expect`
when checking that each test has passed. `assert` is still used for
validating test pre-conditions.

This migration was primarily done to make migrating to TypeScript
easier in the future.