Set password sooner to avoid redundant persistance
The vault password is now always set in the method responsible for
creating the vault. Previously it was set when the vault was persisted,
or when the first keyring was created. This led to redundant calls to
persist the keyrings, just to achieve the side-effect of setting the
password.

This change eliminates those redundant calls, and a few more that had
no obvious purpose.

Note that the password is also set on unlock, and that has not changed
here.