Update ESLint config to v11
The ESLint config has been updated to v11, and all lint-related
packages have been updated.

Most changes were fixed using `yarn lint:fix`. A manual change was
needed to ignore the error for use of the `in` operator, and to ignore
the lint warning for the use of the `events` API. A note was added to
the README to explain that the `events` API is used.

The `lint:misc` script was updated to include a new flag, to better
match the module template.