Fix `@metamask/eth-sig-util` types
The types for the package `@metamask/eth-sig-util` were being
overwritten by a local type declaration. That has been removed, and all
type errors have been resolved.

All of the type errors were related to `normalize` returning
`undefined` if given a nullish input. We never pass it a nullish input,
but TypeScript wasn't able to infer that, so that issue was addressed
with type casts. This isn't ideal, but it's temporary. Once we finish
migrating all Keyrings to the new Keyring interface, we won't need to
normalize addresses anymore at all.