Narrow return type of `signTypedMessage`
The method `signTypedMessage` always returns a string, but the declared
return type was `Bytes` (which is a type union of `string` along with
other things). This made the method more difficult to use, as callers
had to handle the other `Bytes` types as return values as well.

The method has been updated to use `string` as the return type.