Fix `signTypedMessage` parameter types (#250)
The `signTypedMessage` parameter types accepted the data to sign as an
array. This works with `signTypedData_v1`, but is incompatible with
later versions.

The signature has been updated to accept data as either an array or an
object. Additionally, tests have been added to cover all three
supported versions of `signTypedData`.