Change expected options when adding simple keyring
The method `addNewKeying` now expects the second parameter (`opts`) to
be an array of private keys rather than an object with a `privateKeys`
property. This is how it worked prior to v11.

This change was made to reduce the number of changes needed to use this
package. v11 hasn't yet been used on either client or on core, so
undoing this API change made in v11 reduces the work required to update.

This also simplifies the method, removing a special case that was added
just for the simple keyring.