Restore compatibility with QR Keyring
The `KeyringController` method `addNewKeyring` was changed in v11 to
initialize the keyring options to an empty object. These options are
passed to the `deserialize` method of the keyring. This broke
compatibility with the QR keyring because its `deserialize` method does
not expect to be passed an empty object.

The `KeyringController` has been updated to no longer initialize the
`addNewKeyring` options to an empty object. It should now work the same
as it did prior to v11. The one functional difference is that it now
throws an error when `addNewKeyring` is used to add a simple keyring
without supplying any private keys, however that path would have
already thrown an error.