Allow non-object keyring serialized state (#253)
The `addNewKeyring` method has been updated to support non-object
keyring serialized state, which had been supported already on versions
prior to v11.

This allows us to remove a special-case for the simple keyring, which
requires a JSON array as its serialized keyring state.

When constructing a simple keyring, the method `addNewKeying` now
expects the second parameter (`opts`) to be an array of private keys
rather than an object with a `privateKeys` property. This is how it
worked prior to v11.